backend_copy/requirements.txt
aenum==3.1.16
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.19.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
audioop-lts==0.2.1
cartesia==2.0.9
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
coloredlogs==15.0.1
dataclasses-json==0.6.7
decorator==5.2.1
deepgram-sdk==4.7.0
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.17.0
elevenlabs==2.16.0
fastapi==0.109.0
flatbuffers==25.9.23
frozenlist==1.7.0
future==1.0.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
humanfriendly==10.0
idna==3.10
iterators==0.2.0
jiter==0.11.0
joblib==1.5.2
llvmlite==0.44.0
loguru==0.7.3
lxml==6.0.2
Markdown==3.9
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
nltk==3.9.2
numba==0.61.2
numpy==2.2.6
onnxruntime==1.23.0
openai==1.99.1
packaging==25.0
pillow==11.3.0
pipecat-ai==0.0.86
plivo==4.59.2
propcache==0.3.2
protobuf==5.29.5
pydantic==2.11.9
pydantic_core==2.33.2
pydub==0.25.1
PyJWT==2.10.1
pyloudnorm==0.1.1
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
resampy==0.4.3
scipy==1.16.2
six==1.17.0
sniffio==1.3.1
soxr==0.5.0.post1
SQLAlchemy==2.0.43
starlette==0.35.1
sympy==1.14.0
tiktoken==0.11.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.27.0
uvloop==0.21.0
verboselogs==1.7
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
